refactoring upgrade shutdown logic
